New Position Tracking Algorithm
Implemented Odometry using drivetrain motor encoders(tracks rotation of motors, and therefore wheels), an inertial sensor(tracks direction/angle the robot is facing) and trigonometry to calculate the absolute position of the robot on the field as it moves. Added MoveToPoint and TurnToPoint algorithms to be used with odometry, as well as a boomerang controller, basically a movetopoint controller, but with a shifting carrot point to locate the robot at not only a specified position, but also a specified direction.