Code for One World Showcase Event at UC Berkeley
Created new higher scoring autonomous routes to rush towards the middle first, as the game elements in the middle of the field are better to score earlier, due to the ability of the opponents to disturb the middle section. Added minimum output voltage for drivetrain motors during chained movements to maintain speed through separate but chained movements. Implemented maximum acceleration and deceleration limits of the robot going forward and backward to prevent tipping of the bot, due to sudden stopping or braking.