Code for Hollywood Blockbuster Signature Event in Los Angeles
Created new autonomous route for rushing the middle, immune to disruption from opponents, as the bot backs off immediately from the middle after intaking a game element in the middle, which encourages minimal contact from opposing robots in the middle, because the robot stays in the middle for a very short period of time. Added new drive distance linear PID for pushing game elements into the goal, to use as a position resetting device, making odometry more accurate and consistent, as the robot ends up at the same location approximately every time when scoring into the goal, due to the fixed position of the goal.