Vex AI Sensor Test
At the Space City event, we realized that scoring a ring on the colored stake during autonomous wasn't as consistent as expected, because there were slight variations between fields, but scoring on the stake required a low margin of error. Therefore, we decided to try the new Vex AI Sensor, similar to its predecessor the Vision Sensor. Created a new MoveToPoint PID, which used the AI sensor to detect a colored object a move the robot towards it, very consistent at home, but didn't work at the competition when the background was the same color(blue object, blue wall).